+++ src/js/containers/covid19/SummaryInsightsContainer.jsx
@@ -8,7 +8,7 @@ import PropTypes from 'prop-types';
 import { useSelector, useDispatch } from 'react-redux';
 import { setTotals } from 'redux/actions/covid19/covid19Actions';
 import { awardTypeGroups, awardTypeGroupLabels } from 'dataMapping/search/awardType';
-import { fetchAwardAmounts } from 'helpers/disasterHelper';
+import { fetchAwardAmounts } from 'apis/disaster';
 import { InformationBoxes } from "data-transparency-ui";
 import { useInFlightList } from 'helpers/covid19Helper';
 import { isEqual } from 'lodash';
