+++ src/js/containers/covid19/recipient/map/MapContainer.jsx
@@ -27,7 +27,7 @@ import {
 } from 'dataMapping/covid19/recipient/map/map';
 import { awardTypeTabs } from 'dataMapping/covid19/covid19';
 import { awardTypeGroups } from 'dataMapping/search/awardType';
-import { fetchRecipientSpendingByGeography } from 'helpers/disasterHelper';
+import { fetchRecipientSpendingByGeography } from 'apis/disaster';
 import Analytics from 'helpers/analytics/Analytics';
 import SummaryInsightsContainer from '../SummaryInsightsContainer';
